requirements.txt
fastapi
uvicorn
sqlalchemy
aiofiles
python-multipart
pyjwt
bcrypt
